+++ modules/agar/src/main/ts/ephox/agar/pipe/Pipe.ts
@@ -1,5 +1,4 @@
 import { PlatformDetection } from '@ephox/sand';
-import Promise from '@ephox/wrap-promise-polyfill';
 
 import { addStackTrace, TestLogs } from '../api/TestLogs';