+++ modules/tinymce/src/core/main/ts/file/Uploader.ts
@@ -9,7 +9,6 @@ import { Fun, Optional, Type } from '@ephox/katamari';
 
 import { BlobInfo } from '../api/file/BlobCache';
 import { NotificationApi } from '../api/NotificationManager';
-import Promise from '../api/util/Promise';
 import Tools from '../api/util/Tools';
 import { UploadStatus } from './UploadStatus';