+++ modules/tinymce/src/core/main/ts/init/InitContentBody.ts
@@ -24,7 +24,6 @@ import Schema, { SchemaSettings } from '../api/html/Schema';
 import * as Options from '../api/Options';
 import UndoManager from '../api/UndoManager';
 import Delay from '../api/util/Delay';
-import Promise from '../api/util/Promise';
 import Tools from '../api/util/Tools';
 import * as CaretFinder from '../caret/CaretFinder';
 import CaretPosition from '../caret/CaretPosition';