+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/imagetools/Errors.ts
@@ -7,8 +7,6 @@
 
 import { Arr, Fun } from '@ephox/katamari';
 
-import Promise from 'tinymce/core/api/util/Promise';
-
 import * as Utils from './Utils';
 
 const friendlyHttpErrors = [