+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/imagetools/Utils.ts
@@ -5,7 +5,6 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import Promise from 'tinymce/core/api/util/Promise';
 import Tools from 'tinymce/core/api/util/Tools';
 
 const isValue = (obj) => {