+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/autocomplete/AutocompleteTest.ts
@@ -6,7 +6,6 @@ import { SugarBody } from '@ephox/sugar';
 import { TinyAssertions, TinyContentActions, TinyHooks, TinySelections, TinyUiActions } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
-import PromisePolyfill from 'tinymce/core/api/util/Promise';
 
 import { getGreenImageDataUrl } from '../../../module/Assets';
 import {
@@ -38,7 +37,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         ch: '+',
         minChars: 0,
         columns: 1,
-        fetch: (_pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (_pattern, _maxResults) => new Promise((resolve) => {
           resolve(
             Arr.map([ 'aA', 'bB', 'cC', 'dD' ], (letter) => ({
               value: `plus-${letter}`,
@@ -58,7 +57,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         ch: ':',
         minChars: 0,
         columns: 2,
-        fetch: (_pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (_pattern, _maxResults) => new Promise((resolve) => {
           resolve(
             Arr.map([ 'a' ], (letter) => ({
               value: `colon1-${letter}`,
@@ -77,7 +76,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         ch: ':',
         minChars: 0,
         columns: 2,
-        fetch: (_pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (_pattern, _maxResults) => new Promise((resolve) => {
           resolve(
             Arr.map([ 'a', 'b' ], (letter) => ({
               value: `colon2-${letter}`,
@@ -96,7 +95,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         ch: '~',
         minChars: 0,
         columns: 'auto',
-        fetch: (_pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (_pattern, _maxResults) => new Promise((resolve) => {
           resolve(
             Arr.map([ 'a', 'b', 'c', 'd' ], (letter) => ({
               value: `tilde-${letter}`,
@@ -115,7 +114,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         ch: '!',
         minChars: 0,
         columns: 1,
-        fetch: (_pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (_pattern, _maxResults) => new Promise((resolve) => {
           resolve(
             Arr.map([ 'a', 'b', 'c', 'd' ], (letter) => ({
               value: `exclamation-${letter}`,
@@ -136,7 +135,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         matches: (rng, text, _pattern) =>
           // Check the '=' is in the middle of a word
           rng.startOffset !== 0 && !/\s/.test(text.charAt(rng.startOffset - 1)),
-        fetch: (pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (pattern, _maxResults) => new Promise((resolve) => {
           const filteredItems = Arr.filter([ 'two', 'three' ], (number) => number.indexOf(pattern) !== -1);
           resolve(
             Arr.map(filteredItems, (number) => ({
@@ -157,7 +156,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
         ch: '*',
         minChars: 2,
         columns: 'auto',
-        fetch: (_pattern, _maxResults) => new PromisePolyfill((resolve) => {
+        fetch: (_pattern, _maxResults) => new Promise((resolve) => {
           resolve(
             Arr.map([ 'a', 'b', 'c', 'd' ], (letter) => ({
               value: `asterisk-${letter}`,
@@ -184,7 +183,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
             { text: 'equals sign', value: '=' },
             { text: 'some name', value: '`' }
           ], (item) => item.text.indexOf(pattern) !== -1);
-          return new PromisePolyfill((resolve) => {
+          return new Promise((resolve) => {
             resolve(
               Arr.map(filteredItems, (item) => ({
                 value: `hash-${item.value}`,
@@ -211,7 +210,7 @@ describe('browser.tinymce.themes.silver.editor.autocomplete.AutocompleteTest', (
             { text: 'equals sign', value: '=' },
             { text: 'plus sign', value: '+' }
           ], (item) => item.text.indexOf(pattern) !== -1);
-          return new PromisePolyfill((resolve) => {
+          return new Promise((resolve) => {
             resolve(
               Arr.map(filteredItems, (item) => ({
                 value: `euro-${item.value}`,